kf5: Fix build with per-recipe sysroot.
In mose cases this boils down to not hardcoding build paths in output
cmake files. The solution is to use standard OE paths passed to cmake
when it's invoked for each package. Other than that most packages need
to have a few additional deps which previously simply happened to be
present in global sysroot.

Signed-off-by: Piotr Tworek <tworaz@tworaz.net>