kio-extras: disable nfs / enable taglib
Something in oe-coe broke nfs detecttion. It seems that some parts of libtirpc
were deprecated in glibc. A patch was applied to add headers but it linking
fails due to missing symbols (first during configure and if we hack away
symbol detection compile fails)

So disable nfs by patch. On the way we made taglib detection work.

[1] https://git.openembedded.org/openembedded-core/commit/?id=674596421320de08142e010fdd65ec6f0a0f34e9

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>